# tests/api-docs.json
{"openapi":"3.0.1","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/vaga":{"get":{"tags":["vagas-controller"],"operationId":"listar","parameters":[{"name":"paginacao","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDadosDetalhamentoVaga"}}}}}},"put":{"tags":["vagas-controller"],"operationId":"atualizar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAtualizacaoVaga"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoVaga"}}}}}},"post":{"tags":["vagas-controller"],"operationId":"cadastrar","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCadastroVaga"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoVaga"}}}}}}},"/empresa":{"get":{"tags":["empresa-controller"],"operationId":"listar_1","parameters":[{"name":"paginacao","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDadosListagemEmpresa"}}}}}},"put":{"tags":["empresa-controller"],"operationId":"atualizar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAtualizacaoEmpresa"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoEmpresa"}}}}}},"post":{"tags":["empresa-controller"],"operationId":"cadastrar_1","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCadastroEmpresa"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoEmpresa"}}}}}}},"/curriculo":{"get":{"tags":["candidato-controller"],"operationId":"listar_2","parameters":[{"name":"paginacao","in":"query","required":true,"schema":{"$ref":"#/components/schemas/Pageable"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/PageDadosExibirCurriculo"}}}}}},"put":{"tags":["candidato-controller"],"operationId":"atualizar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosAtualizacaoCurriculo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoCurriculo"}}}}}},"post":{"tags":["candidato-controller"],"operationId":"cadastrar_2","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DadosCurriculo"}}},"required":true},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoCurriculo"}}}}}}},"/empresa/{id}":{"get":{"tags":["empresa-controller"],"operationId":"buscar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosDetalhamentoEmpresa"}}}}}},"delete":{"tags":["empresa-controller"],"operationId":"desativar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/curriculo/{id}":{"get":{"tags":["candidato-controller"],"operationId":"buscar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/DadosExibirCurriculo"}}}}}},"delete":{"tags":["candidato-controller"],"operationId":"desativar_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/vaga/{id}":{"delete":{"tags":["vagas-controller"],"operationId":"pausar","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/vaga/remover/{id}":{"delete":{"tags":["vagas-controller"],"operationId":"remover","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/empresa/remover/{id}":{"delete":{"tags":["empresa-controller"],"operationId":"remover_1","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}},"/curriculo/remover/{id}":{"delete":{"tags":["candidato-controller"],"operationId":"remover_2","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK"}}}}},"components":{"schemas":{"DadosAtualizacaoVaga":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"cargo":{"type":"string"},"formatoDeTrabalho":{"type":"string","enum":["REMOTO","HIBRIDO","PRESENCIAL"]},"areaDeAtuacao":{"type":"string"},"cidade":{"type":"string"},"formacaoRequerida":{"type":"string"},"experienciaRequerida":{"type":"integer","format":"int32"},"habilidadesRequeridas":{"type":"string"},"descricao":{"type":"string"},"salario":{"type":"number"}}},"DadosDetalhamentoVaga":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"titulo":{"type":"string"},"cargo":{"type":"string"},"formatoDeTrabalho":{"type":"string","enum":["REMOTO","HIBRIDO","PRESENCIAL"]},"nomeEmpresa":{"type":"string"},"cidade":{"type":"string"},"formacaoRequerida":{"type":"string"},"experienciaRequerida":{"type":"integer","format":"int32"},"habilidadesRequeridas":{"type":"string"},"descricao":{"type":"string"},"salario":{"type":"number"},"dataPublicacao":{"type":"string","format":"date"}}},"DadosAtualizacaoEmpresa":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"razaoSocial":{"type":"string"},"nomeFantasia":{"type":"string"},"contato":{"$ref":"#/components/schemas/DadosContato"},"endereco":{"$ref":"#/components/schemas/DadosEndereco"}}},"DadosContato":{"required":["email"],"type":"object","properties":{"telefone":{"type":"string"},"celular":{"type":"string"},"email":{"type":"string"},"site":{"type":"string"}}},"DadosEndereco":{"required":["bairro","cep","cidade","logradouro","uf"],"type":"object","properties":{"cep":{"pattern":"\\d{8}","type":"string"},"logradouro":{"type":"string"},"numero":{"type":"string"},"complemento":{"type":"string"},"bairro":{"type":"string"},"cidade":{"type":"string"},"uf":{"type":"string"}}},"Contato":{"type":"object","properties":{"telefone":{"type":"string"},"celular":{"type":"string"},"email":{"type":"string"},"site":{"type":"string"}}},"DadosDetalhamentoEmpresa":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"cnpj":{"type":"string"},"razaoSocial":{"type":"string"},"nomeFantasia":{"type":"string"},"contato":{"$ref":"#/components/schemas/Contato"},"endereco":{"$ref":"#/components/schemas/Endereco"}}},"Endereco":{"type":"object","properties":{"cep":{"type":"string"},"logradouro":{"type":"string"},"numero":{"type":"string"},"complemento":{"type":"string"},"bairro":{"type":"string"},"cidade":{"type":"string"},"uf":{"type":"string"}}},"DadosAtualizacaoCurriculo":{"required":["id"],"type":"object","properties":{"id":{"type":"integer","format":"int64"},"contato":{"$ref":"#/components/schemas/DadosContato"},"endereco":{"$ref":"#/components/schemas/DadosEndereco"},"formacao":{"$ref":"#/components/schemas/DadosFormacao"},"dadosProfissionais":{"$ref":"#/components/schemas/DadosProfissionaisCadastro"}}},"DadosFormacao":{"type":"object","properties":{"curso":{"type":"string"},"nivel":{"type":"string"},"instituicao":{"type":"string"},"dataInicio":{"type":"string","format":"date"},"dataConclusao":{"type":"string","format":"date"}}},"DadosProfissionaisCadastro":{"type":"object","properties":{"empresa":{"type":"string"},"cargo":{"type":"string"},"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"},"descricao":{"type":"string"}}},"DadosDetalhamentoCurriculo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"dadosPessoais":{"$ref":"#/components/schemas/DadosPessoais"},"contato":{"$ref":"#/components/schemas/Contato"},"endereco":{"$ref":"#/components/schemas/Endereco"},"formacao":{"$ref":"#/components/schemas/Formacao"},"dadosProfissionais":{"type":"array","items":{"$ref":"#/components/schemas/DadosProfissionais"}}}},"DadosPessoais":{"type":"object","properties":{"nome":{"type":"string"},"sobrenome":{"type":"string"},"cpf":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"genero":{"type":"string"},"nacionalidade":{"type":"string"}}},"DadosProfissionais":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"empresa":{"type":"string"},"cargo":{"type":"string"},"dataInicio":{"type":"string","format":"date"},"dataFim":{"type":"string","format":"date"},"descricao":{"type":"string"}}},"Formacao":{"type":"object","properties":{"curso":{"type":"string"},"nivel":{"type":"string"},"instituicao":{"type":"string"},"dataInicio":{"type":"string","format":"date"},"dataConclusao":{"type":"string","format":"date"}}},"DadosCadastroVaga":{"required":["titulo"],"type":"object","properties":{"titulo":{"type":"string"},"cargo":{"type":"string"},"formatoDeTrabalho":{"type":"string","enum":["REMOTO","HIBRIDO","PRESENCIAL"]},"idEmpresa":{"type":"integer","format":"int64"},"formacaoRequerida":{"type":"string"},"experienciaRequerida":{"type":"integer","format":"int32"},"habilidadesRequeridas":{"type":"string"},"descricao":{"type":"string"},"salario":{"type":"number"},"dataPublicacao":{"type":"string","format":"date"}}},"DadosCadastroEmpresa":{"type":"object","properties":{"empresa":{"$ref":"#/components/schemas/DadosEmpresa"},"contato":{"$ref":"#/components/schemas/DadosContato"},"endereco":{"$ref":"#/components/schemas/DadosEndereco"}}},"DadosEmpresa":{"required":["cnpj","razaoSocial"],"type":"object","properties":{"cnpj":{"type":"string"},"razaoSocial":{"type":"string"},"nomeFantasia":{"type":"string"},"areaDeAtuacao":{"type":"string"}}},"DadosCurriculo":{"type":"object","properties":{"dadosPessoais":{"$ref":"#/components/schemas/DadosPessoaisCadastro"},"contato":{"$ref":"#/components/schemas/DadosContato"},"endereco":{"$ref":"#/components/schemas/DadosEndereco"},"formacao":{"$ref":"#/components/schemas/DadosFormacao"},"dadosProfissionais":{"type":"array","items":{"$ref":"#/components/schemas/DadosProfissionaisCadastro"}}}},"DadosPessoaisCadastro":{"required":["cpf","dataNascimento","nome","sobrenome"],"type":"object","properties":{"nome":{"type":"string"},"sobrenome":{"type":"string"},"cpf":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"genero":{"type":"string"},"nacionalidade":{"type":"string"}}},"Pageable":{"type":"object","properties":{"page":{"minimum":0,"type":"integer","format":"int32"},"size":{"minimum":1,"type":"integer","format":"int32"},"sort":{"type":"array","items":{"type":"string"}}}},"PageDadosDetalhamentoVaga":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DadosDetalhamentoVaga"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"PageableObject":{"type":"object","properties":{"pageNumber":{"type":"integer","format":"int32"},"pageSize":{"type":"integer","format":"int32"},"paged":{"type":"boolean"},"unpaged":{"type":"boolean"},"offset":{"type":"integer","format":"int64"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}}}},"SortObject":{"type":"object","properties":{"direction":{"type":"string"},"nullHandling":{"type":"string"},"ascending":{"type":"boolean"},"property":{"type":"string"},"ignoreCase":{"type":"boolean"}}},"DadosListagemEmpresa":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"razaoSocial":{"type":"string"},"nomeFantasia":{"type":"string"},"contato":{"$ref":"#/components/schemas/Contato"},"endereco":{"$ref":"#/components/schemas/Endereco"}}},"PageDadosListagemEmpresa":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DadosListagemEmpresa"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}},"DadosExibirCurriculo":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"nome":{"type":"string"},"sobrenome":{"type":"string"},"genero":{"type":"string"},"dataNascimento":{"type":"string","format":"date"},"nacionalidade":{"type":"string"},"contato":{"$ref":"#/components/schemas/Contato"},"endereco":{"$ref":"#/components/schemas/Endereco"},"formacao":{"$ref":"#/components/schemas/Formacao"},"dadosProfissionais":{"type":"array","items":{"$ref":"#/components/schemas/DadosProfissionais"}}}},"PageDadosExibirCurriculo":{"type":"object","properties":{"totalPages":{"type":"integer","format":"int32"},"totalElements":{"type":"integer","format":"int64"},"pageable":{"$ref":"#/components/schemas/PageableObject"},"first":{"type":"boolean"},"last":{"type":"boolean"},"size":{"type":"integer","format":"int32"},"content":{"type":"array","items":{"$ref":"#/components/schemas/DadosExibirCurriculo"}},"number":{"type":"integer","format":"int32"},"sort":{"type":"array","items":{"$ref":"#/components/schemas/SortObject"}},"numberOfElements":{"type":"integer","format":"int32"},"empty":{"type":"boolean"}}}}}}